shorten precomp filenames
to 6 hex digits for each part. the few times
i've seen this in the routines list it was just
the filename of the precompiled binary, i.e. MOARVM
blobs. They aren't really interesting to the user
anyway.

In the future, hopefully the profiler backend will
be able to query the precomp repository for what
exactly the file was for.