RoutinePaths accepts callsite ids; use in alloc viewer
since not every version of a given routine will be
allocating the types we're considering, we can't just
use the routine-id version of the RoutinePaths feature.

Instead, we carry around the call ids of the allocating
calls and get a path list/tree for them.

Now you can expand the routines for a type and get a call
tree for it, thus letting you go to the call graph explorer.