add a search function to children-of-call
this is used by the call-graph to get the list
of calls you can click on, and it will allow
the call graph tool to have a little search
input field that lets you see what paths
lead to any routine with a matching name.

Later on, more complicated searches should be
offered.