attempt at making search index creation faster
create an index for parent_id, which is already worth
a whole lot, but for big data sets, like what i get
from profiling "zef list" with about 4 million nodes,
it's still unacceptably slow.

Doing each tree in the forest individually - and hoping
that limiting the ID space that way - doesn't seem to
be a noticeable win, or at least not a factor of 100
that I'd like to have.

Nevertheless, I wanted to commit this to git so I can
look back later and laugh at my naivité.